requirements.txt
djangorestframework
Django>=3.0,<4.0
django-storages==1.8
psycopg2-binary>=2.8
celery
redis
markdown
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-filter
graphene-django
